fix bug in multilabel mode in QRF. added option to get comment in lut prepare

+++ Internal/QRF/QRF/QRF.h
@@ -89,7 +89,7 @@ public:
 	vector<ValInd> qy;				// holding and sorting y values in a node when searching a split
 	vector<int> inds;
 
-	vector<float> histr_sum;
+	vector<double> histr_sum;
 	vector<int>	histr_num;
